pyproject.toml
[project]
name = "mcp-blockchain-rail"
version = "0.1.0"
description = "MCP Server for Blockchain Rail interactions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "web3",
    "requests",
    "pyyaml>=6.0",
    "cryptography>=41.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "types-PyYAML>=6.0"
]

[project.scripts]
mcp-blockchain-rail = "mcp_blockchain_rail.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_blockchain_rail"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["mcp_blockchain_rail"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "tests/",
    ".venv/",
    "build/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
